requirements.txt
pytube==12.1.0
python-telegram-bot==13.13